code refactoring

+++ service/src/main/java/com/senla/library/service/impl/BookServiceImpl.java
@@ -1,17 +1,17 @@
 package com.senla.library.service.impl;
+import com.senla.library.dto.AuthorDTO;
 import com.senla.library.mapper.BookMapper;
 import com.senla.library.repository.BookRepository;
 import com.senla.library.entity.Book;
 import com.senla.library.dto.BookDTO;
 import com.senla.library.service.BookService;
+import com.senla.library.service.exception.ResourceDuplicationException;
+import com.senla.library.service.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -31,28 +31,49 @@ public class BookServiceImpl implements BookService {
     
     @Override
     public BookDTO save(BookDTO dto) {
-        return mapper.toDto(repository.save(mapper.toEntity(dto)));
+        Book book = mapper.toEntity(dto);
+        if(repository.existsByIsbn(book.getIsbn())){
+            throw new ResourceDuplicationException("CONFLICT ISBN, error saving data, " +
+                    "the database contains such data");
+        }else{
+            return mapper.toDto(repository.save(book));
+        }
     }
     
     @Override
     public void save(List<BookDTO> dtos) {
-        repository.saveAll(mapper.toEntityList(dtos));
+        List<Book> books = mapper.toEntityList(dtos);
+        books.forEach(book -> {
+            if(repository.existsByIsbn(book.getIsbn())){
+                throw  new ResourceDuplicationException("CONFLICT, error saving data, " + "the database contains such " +
+                        "data" + book.toString());
+            }
+        });
     }
     
     @Override
     public void deleteById(Long id) {
-        repository.deleteById(id);
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("Failed to delete by primary key ");
+        }
     }
     
     @Override
-    public Optional<BookDTO> findById(Long id) {
-        Optional<Book> entityOptional = repository.findById(id);
-      return entityOptional.map(entity -> Optional.ofNullable(mapper.toDto(entity)).get());
+    public BookDTO findById(Long id) {
+      return mapper.toDto(repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Book by id not found")));
     }
     
     @Override
     public List<BookDTO> findAll() {
-        return mapper.toDtoList((List<Book>) repository.findAll());
+        Iterable<Book> all = repository.findAll();
+        List<Book>books=(List<Book>)all;
+        if(books.isEmpty()){
+            throw new ResourceNotFoundException("Book not found");
+        }else {
+            return mapper.toDtoList(books);
+        }
     }
     
     @Override
@@ -66,7 +87,8 @@ public class BookServiceImpl implements BookService {
     public BookDTO updateById(BookDTO dto) {
         if(repository.existsById(dto.getId())){
             return save(dto);
+        } else {
+            throw new ResourceNotFoundException("update failed no record with this id");
         }
-        return null;
     }
 }
